+++ engine/language-server/src/main/scala/org/enso/languageserver/runtime/ContextRegistry.scala
@@ -10,10 +10,10 @@ import org.enso.languageserver.event.{
 import org.enso.languageserver.monitoring.MonitoringProtocol.{Ping, Pong}
 import org.enso.languageserver.runtime.handler._
 import org.enso.languageserver.util.UnhandledLogging
-import org.enso.logger.MethodsSampler
 import org.enso.logger.akka.ActorMessageLogging
 import org.enso.polyglot.runtime.Runtime.Api
 import org.enso.polyglot.runtime.Runtime.Api.ContextId
+import org.enso.profiling.MethodsSampler
 import org.enso.searcher.SuggestionsRepo
 
 import java.util.UUID
